Trailer effects capture changes.
Trailer effects are now captured using a SpawnNotify, removing the iterator in the Projectile capturer.

Additionally, trailers are accurately moved forward by simply calling 'AutonomousPhysics' (calls physTrailer which is the real PHYS_Trailer placer)

On top of that, ShieldBelt effects and other attachments added to players/pawns are captured as well, fixing the mismatching positions.